refactor (tests): move testdb engine to tests/ folder
